requirements-deploy.txt
setuptools
wheel
twine
packaging
pyclean
pytest
black
isort
flake8
dbx
spetlr-tools